Add previous behaviour to docker container
The docker container continues to work as before if no CRON variable is set.
If execution parameters ere defined through the OPTIONS variable, the container will use those in conujunction with the directories mounted to /mnt/input and /mnt/output. Both of those mount point are declared in the dockerfile.