See Issue https://github.com/ivandokov/phockup/issues/79 (#154)
Implemented --skip-unknown flag to leave files without matching EXIF data in place.
Added test_skip_unknown test to the harness

This has proved useful to me when the folder structure contains metadata that I can use to process images on the second or third pass with additional flags.  Having them moved to a single folder loses the taxonomy that the images may have, so having them stay in place is useful to me.