Add Cronjob and Environment Variables to Docker Image (#147)
* Add Cronjob and environment variables to docker image

* Prevent parallel execution of phockup

* Add clean up lock on container restart

if the container is killed in the middle of a phockup process and the lockfile persists, this will remove the lockfile once before the cron job is set up upon restarting the container.

* Add previous behaviour to docker container

The docker container continues to work as before if no CRON variable is set.
If execution parameters ere defined through the OPTIONS variable, the container will use those in conujunction with the directories mounted to /mnt/input and /mnt/output. Both of those mount point are declared in the dockerfile.

* Change ENV variables to only be considered for execution when running as cron job.

* Update readme with description for running phocker as a cron job.

* Fix naming in readme

I mixed the wrong words there - its photos and back up not photos and docker.