Add support for live photos
Look for MediaGroupUUID and ContentIdentifier EXIF tags. These are
present in media files which are part of live photo pairs. If found, any
future files with the same UUID are named the same as the first one and
placed in the same directory.

With this change, live photos and their associated .mov files are moved
together.