Handle arbitrary sidecar files
Currently support exists for handling .xmp sidecars. This change adds
support for arbitrary sidecar extensions. It also expands the set of
sidecars which are handled by default. The new default set, and the
reason this set was chosen, is:

 .xmp - general metadata format for all image types
 .json - generated by google takeout when exporting from g photos
 .yml/.yaml - used by Photoprism

Users can override this default set by passing the new --sidecars
option. For example, to handle .xyz, users should pass

  --sidecars="xmp,json,yml,xyz"